Combine DSR policy setup scripts together
